backend/requirements.txt
Django==3.0.4
psycopg2-binary==2.8.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
django-filter==2.2.0
drf-yasg==1.17.1
pdbpp==0.10.2
factory-boy==2.12.0
flake8==3.7.9
mypy==0.770
coverage==5.1
Pillow==7.1.2

